7-3 打war包部署到外置tomcat容器

+++ src/main/java/com/kimgao/bootlauch/generator/ArticleMapper.java
@@ -10,7 +10,7 @@ public interface ArticleMapper {
 
     int deleteByExample(ArticleExample example);
 
-    int deleteByPrimaryKey(Long id);
+    int deleteByPrimaryKey(Integer id);
 
     int insert(Article record);
 
@@ -18,7 +18,7 @@ public interface ArticleMapper {
 
     List<Article> selectByExample(ArticleExample example);
 
-    Article selectByPrimaryKey(Long id);
+    Article selectByPrimaryKey(Integer id);
 
     int updateByExampleSelective(@Param("record") Article record, @Param("example") ArticleExample example);
